+++ app/Http/Requests/Discount/StoreDiscountRequest.php
@@ -1,22 +1,17 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Discount;
 
 use Illuminate\Foundation\Http\FormRequest;
 
 class StoreDiscountRequest extends FormRequest
 {
-    /**
-     * Determine if the user is authorized to make this request.
-     */
     public function authorize(): bool
     {
         return false;
     }
 
     /**
-     * Get the validation rules that apply to the request.
-     *
      * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
      */
     public function rules(): array
