+++ app/Http/Resources/DiscountResource.php
@@ -8,8 +8,6 @@ use Illuminate\Http\Resources\Json\JsonResource;
 class DiscountResource extends JsonResource
 {
     /**
-     * Transform the resource into an array.
-     *
      * @return array<string, mixed>
      */
     public function toArray(Request $request): array
