+++ app/Policies/DiscountPolicy.php
@@ -8,59 +8,39 @@ use Illuminate\Auth\Access\Response;
 
 class DiscountPolicy
 {
-    /**
-     * Determine whether the user can view any models.
-     */
+
     public function viewAny(User $user): bool
     {
         return false;
     }
 
-    /**
-     * Determine whether the user can view the model.
-     */
     public function view(User $user, Discount $discount): bool
     {
         return false;
     }
 
-    /**
-     * Determine whether the user can create models.
-     */
     public function create(User $user): bool
     {
         return false;
     }
 
-    /**
-     * Determine whether the user can update the model.
-     */
     public function update(User $user, Discount $discount): bool
     {
         return false;
     }
 
-    /**
-     * Determine whether the user can delete the model.
-     */
-    public function delete(User $user, Discount $discount): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can restore the model.
-     */
-    public function restore(User $user, Discount $discount): bool
-    {
-        return false;
-    }
-
-    /**
-     * Determine whether the user can permanently delete the model.
-     */
     public function forceDelete(User $user, Discount $discount): bool
     {
         return false;
     }
+
+    // public function delete(User $user, Discount $discount): bool
+    // {
+    //     return false;
+    // }
+
+    // public function restore(User $user, Discount $discount): bool
+    // {
+    //     return false;
+    // }
 }
