+++ database/factories/DiscountFactory.php
@@ -10,8 +10,6 @@ use Illuminate\Database\Eloquent\Factories\Factory;
 class DiscountFactory extends Factory
 {
     /**
-     * Define the model's default state.
-     *
      * @return array<string, mixed>
      */
     public function definition(): array
