+++ tests/Feature/Catalog/Discount/DiscountControllerTest.php
@@ -3,14 +3,11 @@
 namespace Tests\Feature\Catalog\Discount;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 
 class DiscountControllerTest extends TestCase
 {
-    /**
-     * A basic feature test example.
-     */
+    use RefreshDatabase;
     public function test_example(): void
     {
         $response = $this->get('/');
