+++ tests/Unit/Models/DiscountTest.php
@@ -6,9 +6,6 @@ use PHPUnit\Framework\TestCase;
 
 class DiscountTest extends TestCase
 {
-    /**
-     * A basic unit test example.
-     */
     public function test_example(): void
     {
         $this->assertTrue(true);
